Move config to repository
