+++ app/Providers/AppServiceProvider.php
@@ -2,7 +2,7 @@
 
 namespace App\Providers;
 
-use App\Helpers\Configuration;
+use App\Repositories\ConfigRepository;
 use App\Repositories\SpinupWpRepository;
 use DeliciousBrains\SpinupWp\SpinupWp;
 use Illuminate\Support\ServiceProvider;
@@ -27,17 +27,19 @@ class AppServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->singleton(SpinupWpRepository::class, function() {
-            return new SpinupWpRepository(new SpinupWp());
+            return new SpinupWpRepository(
+                resolve(SpinupWp::class),
+            );
         });
 
-        $this->app->singleton(Configuration::class, function () {
+        $this->app->singleton(ConfigRepository::class, function () {
             $path = isset($_ENV['APP_ENV']) && $_ENV['APP_ENV'] === 'testing'
                 ? base_path('tests')
                 : ($_SERVER['HOME'] ?? $_SERVER['USERPROFILE']);
 
             $path .= '/.spinupwp/';
 
-            return new Configuration($path);
+            return new ConfigRepository($path);
         });
     }
 }
