+++ tests/Pest.php
@@ -1,10 +1,8 @@
 <?php
 
-use App\Helpers\Configuration;
+use App\Repositories\ConfigRepository;
 use App\Repositories\SpinupWpRepository;
-use DeliciousBrains\SpinupWp\SpinupWp;
 use GuzzleHttp\Client;
-use Illuminate\Support\Facades\Config;
 use LaravelZero\Framework\Testing\TestCase;
 use Tests\CreatesApplication;
 
@@ -55,7 +53,7 @@ expect()->extend('toBeOne', function () {
 
 function setTestConfigFile($profileData = [])
 {
-    $config = resolve(Configuration::class);
+    $config = resolve(ConfigRepository::class);
     file_put_contents($config->configFilePath(), json_encode([
         'default' => array_merge([
             'api_token' => 'myapikey123',
@@ -66,7 +64,7 @@ function setTestConfigFile($profileData = [])
 
 function deleteTestConfigFile($test = '')
 {
-    $configFile = (resolve(Configuration::class))->configFilePath();
+    $configFile = (resolve(ConfigRepository::class))->configFilePath();
     if (!file_exists($configFile)) {
         return;
     }
